expose nan converter
